fix: perbaikan bug stok/uang, kuantitas integer, dan konsolidasi role 8→6
Perbaikan bug kritis:
- StockService: ganti clamp max(0,..) yang membuat oversell lolos diam-diam
  menjadi InsufficientStockException; tambah lockForUpdate pada penyesuaian stok
- POS Hendhys: tambah pengecekan stok yang selama ini tidak ada sama sekali,
  hitung ulang harga & total di sisi server (sebelumnya percaya kiriman klien)
- POS Hendhys & Jihans: jumlahkan kebutuhan per produk agar produk yang sama di
  dua baris keranjang tidak lolos pengecekan stok
- Pending: hapus hanya setelah penjualan tersimpan (dulu dihapus di depan,
  keranjang bisa hilang permanen bila kasir menyegarkan halaman)
- Penerimaan barang: delta stok sadar kondisi baik/rusak + sinkronisasi ulang PO
- PO: cegah produk ganda (distinct) dan alokasikan qty diterima per baris
- Perbaikan SSRF pada image_url produk, null-check cabang, otorisasi transfer
- Tambah fitur void transaksi (stok kembali, status cancelled, audit log)

Kuantitas stok wajib integer (hanya uang yang desimal):
- 41 kolom kuantitas decimal -> integer, 20 aturan validasi numeric -> integer,
  cast model, dan 100 cast (float) -> (int) pada Resources/Controllers

Konsolidasi role 8 -> 6:
- admin_gudang + super_admin_hendhys + super_admin_jihans -> super_admin
- super_admin tidak lagi mengoperasikan POS (kasir saja)
- RolePermissionSeeder diselaraskan ke 6 role

Penamaan: "Hendhys Pusat" -> "Gudang Hendhys" di seluruh label UI.

Test: 137 lulus, termasuk RoleAccessMatrixTest (6 role x 9 rute) dan 21 test
regresi yang diverifikasi lewat mutation testing.

+++ app/Http/Controllers/Gudang/TransferRequestController.php
@@ -47,7 +47,7 @@ class TransferRequestController extends Controller
         $stocks = JihansGudangStock::whereIn('product_id', $transferRequest->details->pluck('product_id'))
             ->pluck('quantity', 'product_id');
         $transferRequest->details->each(function ($detail) use ($stocks) {
-            $detail->warehouse_stock = (float) ($stocks[$detail->product_id] ?? 0);
+            $detail->warehouse_stock = (int) ($stocks[$detail->product_id] ?? 0);
         });
 
         return Inertia::render('Gudang/TransferRequests/Show', [
@@ -71,7 +71,7 @@ class TransferRequestController extends Controller
                 }
 
                 // Business rule: approved qty cannot exceed requested qty.
-                if ((float) $item['quantity_approved'] > (float) $detail->quantity_requested) {
+                if ((int) $item['quantity_approved'] > (int) $detail->quantity_requested) {
                     throw ValidationException::withMessages([
                         'items' => "Kuantitas disetujui untuk produk '" . ($detail->product->name ?? 'Barang') . "' tidak boleh melebihi jumlah permintaan (" . floatval($detail->quantity_requested) . ").",
                     ]);
@@ -79,7 +79,7 @@ class TransferRequestController extends Controller
 
                 $detail->update(['quantity_approved' => $item['quantity_approved']]);
 
-                if ((float) $item['quantity_approved'] < (float) $detail->quantity_requested) {
+                if ((int) $item['quantity_approved'] < (int) $detail->quantity_requested) {
                     $allFulfilled = false;
                 }
             }
