+++ app/Http/Resources/Hendhys/HendhysBranchRequestResource.php
@@ -28,8 +28,8 @@ class HendhysBranchRequestResource extends JsonResource
                 'id'                 => $d->id,
                 'product'            => $d->product?->name ?? '-',
                 'product_code'       => $d->product?->code ?? '-',
-                'quantity_requested' => (float) $d->quantity_requested,
-                'quantity_approved'  => $d->quantity_approved !== null ? (float) $d->quantity_approved : null,
+                'quantity_requested' => (int) $d->quantity_requested,
+                'quantity_approved'  => $d->quantity_approved !== null ? (int) $d->quantity_approved : null,
                 'unit'               => $d->unit?->abbreviation ?? 'PCS',
             ])),
