+++ app/Http/Resources/Hendhys/HendhysStockMovementResource.php
@@ -18,7 +18,7 @@ class HendhysStockMovementResource extends JsonResource
             'created_at'   => $this->created_at?->format('d M Y, H:i'),
             'product'      => $this->whenLoaded('product', fn () => $this->product?->name ?? 'Produk Dihapus'),
             'type'         => $this->type,
-            'quantity'     => (float) $this->quantity,
+            'quantity'     => (int) $this->quantity,
             'source'       => $this->source,
             'reference_id' => $this->reference_id,
             'notes'        => $this->notes,
